+++ ng_trajectory/optimizers/matryoshka/main.py
@@ -91,6 +91,7 @@ P.createAdd("grid", "computed by default", list, "X-size and y-size of the grid 
 P.createAdd("plot_mapping", False, bool, "Whether a grid should be mapped onto the track (to show the mapping).", "init (viz.)")
 P.createAdd("save_matryoshka", None, str, "Name of the file to save Matryoshka mapping. When unset, do not save.", "init (Matryoshka)")
 P.createAdd("load_matryoshka", None, str, "Name of the file to load Matryoshka from. When unset, do not load.", "init (Matryoshka)")
+P.createAdd("force_load_matryoshka", False, bool, "Whether the transformation should be loaded every time.", "init (Matryoshka)")
 P.createAdd("plot_group_indices", True, bool, "Whether group indices should be shown on the track.", "init (viz.)")
 P.createAdd("plot_group_borders", True, bool, "Whether group borders should be shown on the track.", "init (viz.)")
 P.createAdd("fixed_segments", [], list, "Points to be used instead their corresponding segment.", "init")
@@ -243,8 +244,18 @@ def init(
 
     P.updateAll(kwargs)
 
-    # Load Matryoshka
-    if MATRYOSHKA is None and P.getValue("load_matryoshka") is not None:
+    # Load Matryoshka if:
+    #  - There is no transformation.
+    #  - Current transformation does not work for selected number of segments.
+    #  - We want to load the transformation (param 'force_load_matryoshka').
+    if (
+        P.getValue("load_matryoshka") is not None
+        and (
+            MATRYOSHKA is None
+            or (groups > 0 and groups != len(MATRYOSHKA))
+            or P.getValue("force_load_matryoshka")
+        )
+    ):
         try:
             _data = numpy.load(
                 P.getValue("load_matryoshka"),
@@ -671,7 +682,11 @@ def _opt(points: numpy.ndarray) -> float:
             logvvv ("pointsT:%s" % str(_points.tolist()), level = VERBOSITY)
             logvv ("penalty:%f" % penalty, level = VERBOSITY)
             logfileFlush()
-        return penalty
+
+        # When not optimizing, continue to plot trajectories as well and
+        # not just penalty points.
+        if CRITERION_ARGS.get("optimization", True):
+            return penalty
 
     _c = CRITERION.compute(
         **{